Fix: Rename ExitManger to ExitManager and update imports
