+++ Frontend/EnglishReadyUI/src/Components/Writing/ExitManager/ExitManager.jsx
