+++ Frontend/EnglishReadyUI/src/Components/Writing/ExitManager/__tests__/ExitManager.test.jsx
